+++ Rev6/CatapultBallExplosion.mm
@@ -14,9 +14,9 @@
 	return [self initWithTotalParticles:7];
 }
 
--(id) initWithTotalParticles:(int)p
-{
-	if( !(self=[super initWithTotalParticles:p]) )
+-(id) initWithTotalParticles:(NSUInteger)p {
+
+	if(!(self = [super initWithTotalParticles:p]))
 		return nil;
 	
 	// duration
